Split out DateFormatControl and update the Comment Date block
